Refactor DO structure in tests
