+++ tests/unit/data/observatory/catalog/examples.py
@@ -1,11 +1,11 @@
-from cartoframes.data.observatory.variable import Variable
-from cartoframes.data.observatory.dataset import CatalogDataset
-from cartoframes.data.observatory.category import Category
-from cartoframes.data.observatory.geography import Geography
-from cartoframes.data.observatory.country import Country
-from cartoframes.data.observatory.provider import Provider
-from cartoframes.data.observatory.variable_group import VariableGroup
-from cartoframes.data.observatory.entity import CatalogList
+from cartoframes.data.observatory.catalog.variable import Variable
+from cartoframes.data.observatory.catalog.dataset import CatalogDataset
+from cartoframes.data.observatory.catalog.category import Category
+from cartoframes.data.observatory.catalog.geography import Geography
+from cartoframes.data.observatory.catalog.country import Country
+from cartoframes.data.observatory.catalog.provider import Provider
+from cartoframes.data.observatory.catalog.variable_group import VariableGroup
+from cartoframes.data.observatory.catalog.entity import CatalogList
 
 db_country1 = {'id': 'esp'}
 db_country2 = {'id': 'usa'}
