+++ tests/unit/data/observatory/catalog/repository/test_category_repo.py
@@ -1,11 +1,10 @@
 import pytest
 
-from cartoframes.data.observatory.category import Category
-
 from cartoframes.exceptions import DiscoveryException
-from cartoframes.data.observatory.entity import CatalogList
-from cartoframes.data.observatory.repository.category_repo import CategoryRepository
-from cartoframes.data.observatory.repository.repo_client import RepoClient
+from cartoframes.data.observatory.catalog.category import Category
+from cartoframes.data.observatory.catalog.entity import CatalogList
+from cartoframes.data.observatory.catalog.repository.category_repo import CategoryRepository
+from cartoframes.data.observatory.catalog.repository.repo_client import RepoClient
 from ..examples import test_category1, test_categories, db_category1, db_category2
 
 try:
