+++ tests/unit/data/observatory/catalog/repository/test_variable_group_repo.py
@@ -1,10 +1,10 @@
 import pytest
 
 from cartoframes.exceptions import DiscoveryException
-from cartoframes.data.observatory.entity import CatalogList
-from cartoframes.data.observatory.variable_group import VariableGroup
-from cartoframes.data.observatory.repository.variable_group_repo import VariableGroupRepository
-from cartoframes.data.observatory.repository.repo_client import RepoClient
+from cartoframes.data.observatory.catalog.entity import CatalogList
+from cartoframes.data.observatory.catalog.variable_group import VariableGroup
+from cartoframes.data.observatory.catalog.repository.variable_group_repo import VariableGroupRepository
+from cartoframes.data.observatory.catalog.repository.repo_client import RepoClient
 from ..examples import test_variable_group1, test_variables_groups, db_variable_group1, db_variable_group2
 
 try:
