+++ tests/unit/data/observatory/catalog/test_category.py
@@ -1,10 +1,10 @@
 import pandas as pd
-from cartoframes.data.observatory.repository.geography_repo import GeographyRepository
 
-from cartoframes.data.observatory.category import Category
-from cartoframes.data.observatory.repository.category_repo import CategoryRepository
-from cartoframes.data.observatory.repository.dataset_repo import DatasetRepository
-from cartoframes.data.observatory.entity import CatalogList
+from cartoframes.data.observatory.catalog.category import Category
+from cartoframes.data.observatory.catalog.repository.dataset_repo import DatasetRepository
+from cartoframes.data.observatory.catalog.repository.category_repo import CategoryRepository
+from cartoframes.data.observatory.catalog.repository.geography_repo import GeographyRepository
+from cartoframes.data.observatory.catalog.entity import CatalogList
 from .examples import test_category1, test_datasets, test_categories, db_category1, test_category2, db_category2, \
     test_geographies
