+++ tests/unit/data/observatory/catalog/test_variable_group.py
@@ -1,9 +1,9 @@
 import pandas as pd
 
-from cartoframes.data.observatory.entity import CatalogList
-from cartoframes.data.observatory.variable_group import VariableGroup
-from cartoframes.data.observatory.repository.variable_repo import VariableRepository
-from cartoframes.data.observatory.repository.variable_group_repo import VariableGroupRepository
+from cartoframes.data.observatory.catalog.entity import CatalogList
+from cartoframes.data.observatory.catalog.variable_group import VariableGroup
+from cartoframes.data.observatory.catalog.repository.variable_repo import VariableRepository
+from cartoframes.data.observatory.catalog.repository.variable_group_repo import VariableGroupRepository
 from .examples import test_variables_groups, test_variable_group1, test_variables, db_variable_group1, \
     test_variable_group2, db_variable_group2
